gateway: split the OpenAI-family payload builders for the module line budget
streaming_requests crossed the 1000-line gate; the native Responses and
OpenAI-compatible Chat builders move to openai_payloads.py, mirroring the
existing Messages-family split (messages_payloads), with
dialect_stream_payload remaining the single dispatch seam. No behavior
change.
