+++ exp/runtime/openai_protocol/requests.py
@@ -9,13 +9,7 @@ from typing import Literal, cast
 from openai.types import EmbeddingCreateParams
 from openai.types.chat.completion_create_params import CompletionCreateParams
 from openai.types.responses.response_create_params import ResponseCreateParams
-from pydantic import (
-    BaseModel,
-    Field,
-    JsonValue,
-    TypeAdapter,
-    ValidationError,
-)
+from pydantic import BaseModel, Field, JsonValue, TypeAdapter, ValidationError
 from pydantic_core import ErrorDetails
 
 from exp.common.core.artifacts import ContractModel, JsonObject
